requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
supabase
pydub
requests
pandas
numpy
google-generativeai
torch
torchaudio
transformers
tqdm
tensorflow
nemo_toolkit[tts]
soundfile
pydub
moviepy